dynamic get component
