+++ web/src/main/gui/web/.roadhogrc.mock.js
@@ -6,18 +6,18 @@ const apis = {
   'POST /public/api/my/customers' :'http://www.potafish.com',
   'POST /public/api/notification/my' : 'http://www.potafish.com',
   'GET /public/api/menu' : {status: 200, data:[
-    {title:'客户管理', content:"TEst",children:[
-      {title:'我的所有客户', key: 1,
+    {title:'客户管理',name:'CUSTOMER',children:[
+      {title:'我的所有客户', id: 1, name: 'MY_CUSTOMER',
         children:[
-          {title:'客户资料', key:1 },
-          {title:'联系人信息', key:2},
-          {title:'跟单记录', key:3},
-          {title:'订单记录', key:4},
-          {title:'操作记录', key:5}]},
-      {title:'共享的客户', key: 2}]},
-    {title:'通知中心', children:[
-      {title: '客户操作通知', key:1},
-      {title: '审核通知', key:2, content:'开发中...'},]},
+          {title:'客户资料', name:'CUSTOMER_INFO', id:1 },
+          {title:'联系人信息', name:'CUSTOMER_CONTACT', id:2},
+          {title:'跟单记录', name:'CUSTOMER_CONTRACT', id:3},
+          {title:'订单记录', id:4},
+          {title:'操作记录', id:5}]},
+      {title:'共享的客户', id: 2 ,name: 'MY_CUSTOMERS', content: 'coding'}]},
+    {title:'通知中心',name:'NOTIFICATION', children:[
+      {title: '客户操作通知', id:1, name: 'CUSTOMER_NOTIFICATION'},
+      {title: '审核通知', id:2, content:'开发中...'},]},
     {title:'跟单管理', content:"维护中...", children:[]},
     {title:'订单管理'},
     {title:'财务管理'},
